revert moving apis
